// view/src/gen/enums/mylist_pb.ts
// @generated by protoc-gen-es v1.10.1 with parameter "target=ts"
// @generated from file enums/mylist.proto (package enums, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { proto3 } from "@bufbuild/protobuf";

/**
 * AttachmentType
 *
 * @generated from enum enums.AttachmentType
 */
export enum AttachmentType {
  /**
   * @generated from enum value: ATTACHMENT_TYPE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: ATTACHMENT_TYPE_PICTURE = 1;
   */
  PICTURE = 1,

  /**
   * @generated from enum value: ATTACHMENT_TYPE_MOVIE = 2;
   */
  MOVIE = 2,
}
// Retrieve enum metadata with: proto3.getEnumType(AttachmentType)
proto3.util.setEnumType(AttachmentType, "enums.AttachmentType", [
  { no: 0, name: "ATTACHMENT_TYPE_UNSPECIFIED" },
  { no: 1, name: "ATTACHMENT_TYPE_PICTURE" },
  { no: 2, name: "ATTACHMENT_TYPE_MOVIE" },
]);

/**
 * ClearType
 *
 * @generated from enum enums.ClearType
 */
export enum ClearType {
  /**
   * @generated from enum value: CLEAR_TYPE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: CLEAR_TYPE_NOT_CLEARED = 1;
   */
  NOT_CLEARED = 1,

  /**
   * @generated from enum value: CLEAR_TYPE_CLEARED = 2;
   */
  CLEARED = 2,

  /**
   * @generated from enum value: CLEAR_TYPE_FULL_COMBO = 3;
   */
  FULL_COMBO = 3,

  /**
   * @generated from enum value: CLEAR_TYPE_ALL_PERFECT = 4;
   */
  ALL_PERFECT = 4,
}
// Retrieve enum metadata with: proto3.getEnumType(ClearType)
proto3.util.setEnumType(ClearType, "enums.ClearType", [
  { no: 0, name: "CLEAR_TYPE_UNSPECIFIED" },
  { no: 1, name: "CLEAR_TYPE_NOT_CLEARED" },
  { no: 2, name: "CLEAR_TYPE_CLEARED" },
  { no: 3, name: "CLEAR_TYPE_FULL_COMBO" },
  { no: 4, name: "CLEAR_TYPE_ALL_PERFECT" },
]);